Refactor presentation layer (#114)
* Initial support for new Teams

* Note delay and what is reported

* Display current system in About dialog

* Use last availability if new one can't be found

* Rename enum to more closely match lingo

* Use a different approach since previous has too much lag

The lag between status change and the associated glyph change being
logged could be tens of minutes.

* Refactor presentation layer

* Bad merge fixes

* Use same project attributes

AssemblyAppInfo gets the executing assembly info. When running the app, that's
TeamsStatusPub project but when running unit tests it's TeamsStatusPub.Core.
Want to have integration tests that actually read info from a project file but
don't want to compile that info into the final build.