Refactor to keep track of whether About window is being shown
Remove view model altogether in favor of using code behind for creating and
showing windows. View model doesn't seem like the right place for that,
especially if it's going to be unit tested.

This also fixes an exception that was thrown when attempting to show the
About window more than once.